% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhapply_functions.R
\name{count_}
\alias{count_}
\title{Count of Most Common Neighbour}
\usage{
count_(mat_open, sub.FUN = "max", count.output = "value")
}
\arguments{
\item{mat_open}{is a 2 dimensional input matrix}

\item{sub.FUN}{describes, if the max or min count should be calculated}

\item{count.output}{specifies if the output should be the value of the
max or min count, or the count itself (number of occurrences)}
}
\description{
The function outputs a vector of the number of times, the most
    common neighbour occurs
}
