% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhapply_functions.R
\name{extreme}
\alias{extreme}
\title{Maximum of Columns}
\usage{
extreme(mat_open, sub.FUN = "max")
}
\arguments{
\item{mat_open}{is a 2 dimensional input matrix}

\item{sub.FUN}{describes, if the max or min value should be calculated}
}
\description{
The function outputs a vector of the max/min values of each column
}
