% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhapply_manipulation.R
\name{generate_coords}
\alias{generate_coords}
\title{Matrix of Neighbourhood to Coordinates}
\usage{
generate_coords(mat)
}
\arguments{
\item{mat}{input is type matrix with odd no. of rows/columns, element in 
middle is pixel.}
}
\description{
The function takes an odd 2D matrix as input. Non zero fields
    are used and value saved. Output is a list of a coordinates matrix and a 
    weights matrix.
}
\examples{
co <- generate_coords(matrix(sample(0:1, 9, rep = TRUE), ncol = 3))
}
