% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhapply_manipulation.R
\name{generate_coords_default}
\alias{generate_coords_default}
\title{Coordinates of Default Pixel-Neighbourhoods}
\usage{
generate_coords_default(neighb_type = 1, width = 1, include.own = FALSE)
}
\arguments{
\item{neighb_type}{specifies the type of surrounding:
Type 1 is square-shaped ("Moore-Neighbourhood")
Type 2 is cross-shaped ("Von-Neumann-Neighbourhood")
Type 3 is diamond-shape}

\item{width}{specifies how many neighbour pixels horizontally and vertically
from center pixel should be included}

\item{include.own}{tells, if the value of the center pixel should be included}
}
\description{
A Function to generate built-in Pixel-Neighbourhood of custom size
}
\examples{
co <- generate_coords_default(neighb_type = "moore", width = 2)
}
