% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhapply_manipulation.R
\name{generate_lags}
\alias{generate_lags}
\title{Generate Matrix Lags}
\usage{
generate_lags(neighb, rows, cols, transition = TRUE)
}
\arguments{
\item{neighb}{is a matrix with two columns each representing the x-axis and
y-axis of a pixel in the pixel neighbourhood.}

\item{rows, cols}{describes how many columns/rows the matrix has.}

\item{transition}{this parameter describes, if the matrix should continue
at the other side of the screen if ending at one.}
}
\description{
This function takes coordinates (2-col matrices) as inputs and 
    generates a list with the dimensions "rows" and "cols" being indices
    on how to lag a matrix to achieve desired output.
}
\examples{
l <- generate_lags(generate_coords_default(2, 2), 10, 10, FALSE)
}
