% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhapply_manipulation.R
\name{open_matrix}
\alias{open_matrix}
\title{Open matrix}
\usage{
open_matrix(mat, lags)
}
\arguments{
\item{mat}{is the matrix to open and must be 2-dimensional.}

\item{lags}{is a list of coordinates describing the pixel neighbourhoods.}
}
\description{
This function takes a matrix and a list of lagged rows/columns 
    as inputs and creates a 2-dimensional matrix with as many rows as objects
     in the neighbourhood.
}
