% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhapply_functions.R
\name{w_sums}
\alias{w_sums}
\title{Sum of Neighbourhood}
\usage{
w_sums(mat_open, weights = 1)
}
\arguments{
\item{mat_open}{is a 2 dimensional input matrix}

\item{weights}{is a vector of weights that is used when summing up a 
neighbourhood. Vector length must be the same as points in neighbourhood.}
}
\description{
The function outputs a the weighted col-sums as a vector.
}
